Delete dispatched lined on order deletion

+++ htdocs/fourn/commande/card.php
@@ -63,6 +63,10 @@ if (isModEnabled('variants')) {
 	require_once DOL_DOCUMENT_ROOT.'/variants/class/ProductCombination.class.php';
 }
 
+if (isModEnabled('stock')) {
+	require_once DOL_DOCUMENT_ROOT.'/fourn/class/fournisseur.commande.dispatch.class.php';
+	require_once DOL_DOCUMENT_ROOT.'/product/stock/class/mouvementstock.class.php';
+}
 
 // Load translation files required by the page
 $langs->loadLangs(array('admin', 'orders', 'sendings', 'companies', 'bills', 'propal', 'receptions', 'supplier_proposal', 'deliveries', 'products', 'stocks', 'productbatch'));
@@ -1100,12 +1104,60 @@ if (empty($reshook)) {
 
 
 	if ($action == 'confirm_delete' && $confirm == 'yes' && $usercandelete) {
-		$result = $object->delete($user);
-		if ($result > 0) {
-			header("Location: ".DOL_URL_ROOT.'/fourn/commande/list.php?restore_lastsearch_values=1');
-			exit;
-		} else {
-			setEventMessages($object->error, $object->errors, 'errors');
+		// Delete existing dispatched lines
+		$dispatchedLines = $object->getDispachedLines();
+		$errOnDelete = 0;
+		if(!empty($dispatchedLines)){
+			foreach($dispatchedLines as $dispatchedLine){
+				$db->begin();
+				$supplierorderdispatch = new CommandeFournisseurDispatch($db);
+				$result = $supplierorderdispatch->fetch($dispatchedLine['id']);
+				if ($result > 0) {
+					$qty = $supplierorderdispatch->qty;
+					$entrepot = $supplierorderdispatch->fk_entrepot;
+					$product = $supplierorderdispatch->fk_product;
+					$price = price2num(GETPOST('price', 'alpha'), 'MU');
+					$comment = $langs->trans('SupplierOrderDeletion', $object->ref);
+					$eatby = $supplierorderdispatch->eatby;
+					$sellby = $supplierorderdispatch->sellby;
+					$batch = $supplierorderdispatch->batch;
+			
+					$result = $supplierorderdispatch->delete($user);
+				}
+				if ($result < 0) {
+					$errorsOnDelete = $object->errors;
+					$errOnDelete++;
+				} else {
+					// If module stock is enabled and the stock increase is done on purchase order dispatching
+					if ($entrepot > 0 && !empty($conf->stock->enabled) && !empty($conf->global->STOCK_CALCULATE_ON_SUPPLIER_DISPATCH_ORDER) && empty($supplierorderdispatch->fk_reception)) {
+						$mouv = new MouvementStock($db);
+						if ($product > 0) {
+							$mouv->origin = &$object;
+							$mouv->setOrigin($object->element, $object->id);
+							$result = $mouv->livraison($user, $product, $entrepot, $qty, $price, $comment, '', $eatby, $sellby, $batch);
+							if ($result < 0) {
+								$errorsOnDelete = $mouv->errors;
+								$errOnDelete++;
+							}
+						}
+					}
+				}
+				if ($errOnDelete > 0) {
+					$db->rollback();
+					setEventMessages('', $errorsOnDelete, 'errors');
+				} else {
+					$db->commit();
+				}
+			}
+		}
+		if(empty($errOnDelete)){
+			$result = $object->delete($user);
+			if ($result > 0) {
+				header("Location: ".DOL_URL_ROOT.'/fourn/commande/list.php?restore_lastsearch_values=1');
+				exit;
+			} else {
+				setEventMessages($object->error, $object->errors, 'errors');
+			}
 		}
 	}
 
@@ -1904,7 +1956,12 @@ if ($action == 'create') {
 
 	// Confirmation de la suppression de la commande
 	if ($action == 'delete') {
-		$formconfirm = $form->formconfirm($_SERVER["PHP_SELF"].'?id='.$object->id, $langs->trans('DeleteOrder'), $langs->trans('ConfirmDeleteOrder'), 'confirm_delete', '', 0, 2);
+		$textmodal = $langs->trans('ConfirmDeleteOrder');
+		if(!empty($object->getDispachedLines())){
+			$textmodal .= "<br/>";
+			$textmodal .= $langs->trans('ExistingDipatchLinesWillBeRemoved');
+		}
+		$formconfirm = $form->formconfirm($_SERVER["PHP_SELF"].'?id='.$object->id, $langs->trans('DeleteOrder'), $textmodal, 'confirm_delete', '', 0, 2);	
 	}
 
 	// Clone confirmation
